Separate auxiliary test types and macros
Allow them to be reused in future tests.
